+++ tests/macho/test_builder.py
@@ -6,6 +6,7 @@ import pathlib
 import os
 import pytest
 import random
+from pathlib import Path
 from subprocess import Popen
 
 import lief
@@ -631,3 +632,20 @@ def test_rpath(tmp_path):
     new_rpaths = list(new.rpaths)
     assert new_rpaths[0].path == "/foo"
     assert new_rpaths[1].path == "/bar"
+
+
+def test_encryption_info(tmp_path: Path):
+    # c.f. https://github.com/lief-project/LIEF/issues/1173
+    macho = lief.MachO.parse(get_sample("MachO/RNCryptor.bin")).at(0)
+    assert macho.encryption_info.crypt_offset != 0
+    macho.encryption_info.crypt_offset = 0
+    macho.encryption_info.crypt_size = 0
+    macho.encryption_info.crypt_id = 0
+
+    output = f"{tmp_path.as_posix()}/new.macho"
+    macho.write(output)
+
+    new = lief.MachO.parse(output).at(0)
+    assert new.encryption_info.crypt_id == 0
+    assert new.encryption_info.crypt_size == 0
+    assert new.encryption_info.crypt_offset == 0